Put firmware name logic in BaseContikiMoteType
ContikiMoteType uses special naming rules, so implement
that as an override in ContikiMoteType itself.
